Add files via upload
Of course I couldn't figure out how to do the code right on here, so I did the code on the Kaggle page, downloaded it and uploaded it here...
I added comments (what I would have written for the report) for each step, so hopefully they are helpful. I wrote very little of the code, but adapted code from others. I added links to the ones I found most helpful on the Readme page.